+++ pyproject.toml
@@ -14,6 +14,9 @@ classifiers = [
 "Homepage" = "https://github.com/WorldofKerry/Python2Verilog/"
 "Bug Tracker" = "https://github.com/WorldofKerry/Python2Verilog/issues"
 
+[project.optional-dependencies]
+visual = ["matplotlib", "numpy", "dash_cytoscape", "dash"]
+
 [tool.setuptools.packages.find]
 "include" = ["python2verilog*"]
