+++ python2verilog/api/exit_handler.py
@@ -12,8 +12,12 @@ from python2verilog import ir
 from python2verilog.api.file_namespaces import file_namespaces
 from python2verilog.api.modes import Modes
 from python2verilog.api.wrappers import context_to_text
-from python2verilog.utils.assertions import assert_typed, assert_typed_dict, get_typed
 from python2verilog.utils.decorator import decorator_with_args
+from python2verilog.utils.smart_asserts import (
+    assert_typed,
+    assert_typed_dict,
+    get_typed,
+)
 
 try:
     # For iPython
