+++ python2verilog/ir/context.py
@@ -17,9 +17,13 @@ from python2verilog.ir.expressions import Var
 from python2verilog.ir.graph import DoneNode
 from python2verilog.ir.instance import Instance
 from python2verilog.ir.signals import ProtocolSignals
-from python2verilog.utils.assertions import assert_typed_dict, get_typed, get_typed_list
 from python2verilog.utils.env_vars import is_debug_mode
 from python2verilog.utils.generics import GenericReprAndStr
+from python2verilog.utils.smart_asserts import (
+    assert_typed_dict,
+    get_typed,
+    get_typed_list,
+)
 
 DEFAULT_STATE_NAME = "___PYTHON_2_VERILOG_STATE___"
