+++ python2verilog/utils/lines.py
