+++ python2verilog/utils/smart_asserts.py
