+++ tests/integration/test_integration.py
@@ -19,7 +19,7 @@ from python2verilog.extern.iverilog import (
     run_iverilog_with_fifos,
     run_iverilog_with_files,
 )
-from python2verilog.utils.assertions import get_typed
+from python2verilog.utils.smart_asserts import get_typed
 from tools.visualization import make_visual
 
 from .cases import TEST_CASES
